requirements.txt
streamlit>=1.32.0
# openai>=1.12.0
python-dotenv>=1.0.0
huggingface_hub==1.7.1
supabase>=2.0.0
